Largely de-duplicates serialization code.
