+++ Server/IO/ISerializable.cs
@@ -1,0 +1,9 @@
+namespace Server
+{
+	public interface ISerializable
+	{
+		int TypeReference { get; }
+		int SerialIdentity { get; }
+		void Serialize(GenericWriter writer);
+	}
+}
