+++ Server/Server.csproj
@@ -142,7 +142,13 @@
     <Compile Include=".\Sector.cs" />
     <Compile Include=".\SecureTrade.cs" />
     <Compile Include=".\Serial.cs" />
-    <Compile Include=".\Serialization.cs" />
+    <Compile Include="EnumerableExtensions.cs" />
+    <Compile Include="IO\AsyncWriter.cs" />
+    <Compile Include="IO\BinaryFileReader.cs" />
+    <Compile Include="IO\BinaryFileWriter.cs" />
+    <Compile Include="IO\GenericWriter.cs" />
+    <Compile Include="IO\ISerializable.cs" />
+    <Compile Include="IO\GenericReader.cs" />
     <Compile Include=".\Skills.cs" />
     <Compile Include=".\Targeting\LandTarget.cs" />
     <Compile Include=".\Targeting\MultiTarget.cs" />
